Refactor navigation to remove bottom tab
