+++ mobile/src/Teacher/index.js
@@ -1,5 +1,5 @@
 /* eslint react/prop-types: 0 */
-
+import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 import { Platform, Image, Text } from 'react-native'
 import { NavigationContainer } from '@react-navigation/native'
@@ -9,42 +9,51 @@ import { colors } from '../utils/theme'
 import { HeaderTitle } from '@react-navigation/stack'
 import DetailScreen from './screens/Explore/DetailScreen'
 
-
 const Tab = createBottomTabNavigator()
+
+const ExploreScreen = () => {
+  return (<Tab.Navigator
+    initialRouteName="Explore"
+    screenOptions={({ route }) => ({
+      tabBarIcon: ({ focused }) => {
+        let iconName
+
+        if (route.name === 'Explore') {
+          iconName = focused ? require('./../assets/images/explore_selected.png') : require('./../assets/images/explore.png')
+        } else if (route.name === 'My Games') {
+          iconName = focused ? require('./../assets/images/mygames_selected.png') : require('./../assets/images/mygames.png')
+        } else if (route.name === 'Quiz Maker') {
+          iconName = focused ? require('./../assets/images/quizmaker_selected.png') : require('./../assets/images/quizmaker.png')
+        } else if (route.name === 'Reports') {
+          iconName = focused ? require('./../assets/images/reports_selected.png') : require('./../assets/images/reports.png')
+        }
+        return <Image source={iconName} />
+      },
+    }), {headerShown: false}}
+    tabBarOptions={{
+      activeTintColor: 'white',
+      inactiveTintColor: '#85C6FF',
+      style: { backgroundColor: '#043272' },
+      tabBarVisible: false
+    }}
+  >
+    <Tab.Screen name="Explore" component={ExploreStackScreen}/>
+    <Tab.Screen name="My Games" component={ExploreStackScreen} />
+    <Tab.Screen name="Quiz Maker" component={ExploreStackScreen} />
+    <Tab.Screen name="Reports" component={ExploreStackScreen} />
+  </Tab.Navigator>)
+}
+
+const ExploreStack = createStackNavigator()
 
 const TeacherApp = ({ props, navigation }) => {
   return (
-    <Tab.Navigator
-      initialRouteName="Explore"
-      screenOptions={({ route }) => ({
-        tabBarIcon: ({ focused }) => {
-          let iconName
-
-          if (route.name === 'Explore') {
-            iconName = focused ? require('./../assets/images/explore_selected.png') : require('./../assets/images/explore.png')
-          } else if (route.name === 'My Games') {
-            iconName = focused ? require('./../assets/images/mygames_selected.png') : require('./../assets/images/mygames.png')
-          } else if (route.name === 'Quiz Maker') {
-            iconName = focused ? require('./../assets/images/quizmaker_selected.png') : require('./../assets/images/quizmaker.png')
-          } else if (route.name === 'Reports') {
-            iconName = focused ? require('./../assets/images/reports_selected.png') : require('./../assets/images/reports.png')
-          }
-          return <Image source={iconName} />
-        },
-      })}
-      tabBarOptions={{
-        activeTintColor: 'white',
-        inactiveTintColor: '#85C6FF',
-        style: { backgroundColor: '#043272' },
-        tabBarVisible: false
-        
-      }}
-    >
-      <Tab.Screen name="Explore" component={ExploreStackScreen}/>
-      <Tab.Screen name="My Games" component={ExploreStackScreen} />
-      <Tab.Screen name="Quiz Maker" component={ExploreStackScreen} />
-      <Tab.Screen name="Reports" component={ExploreStackScreen} />
-    </Tab.Navigator>
+    <ExploreStack.Navigator screenOptions={{
+      headerShown: false
+    }}>
+      <ExploreStack.Screen name="ExploreScreen" component={ExploreScreen} />
+      <ExploreStack.Screen name="GameDetails" component={DetailScreen} />
+    </ExploreStack.Navigator>
   )
 }
